requirements.txt
fastapi==0.124.4
uvicorn[standard]==0.38.0
SQLAlchemy==2.0.45
alembic==1.17.2
PyMySQL==1.1.2
pydantic==2.12.5
pydantic-settings==2.12.0
email-validator==2.3.0
python-dotenv==1.2.1
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose==3.5.0
cryptography==46.0.3